refactor: Container 컴포넌트에서 Background 컴포넌트 분리 (#260)
